move clipboard actions to clipboard dir
