+++ src/modules/editorStore/actions/clipboard/copySelectedColumns.ts
@@ -1,5 +1,5 @@
-import { useTablatureEditorStore } from '../useTablatureEditorStore';
-import { validateColumnSelection } from '../utils/validateColumnSelection';
+import { useTablatureEditorStore } from '../../useTablatureEditorStore';
+import { validateColumnSelection } from '../../utils/validateColumnSelection';
 
 export const copySelectedColumns = () =>
 	useTablatureEditorStore.setState((state) => {
