+++ src/modules/editorStore/actions/clipboard/insertClipboard.ts
@@ -1,5 +1,5 @@
-import { useTablatureEditorStore } from '../useTablatureEditorStore';
-import { insertColumns } from '../utils/insertColumns';
+import { useTablatureEditorStore } from '../../useTablatureEditorStore';
+import { insertColumns } from '../../utils/insertColumns';
 
 export const insertClipboard = () =>
 	useTablatureEditorStore.setState((state) => {
